Create `README.md` from `timezone.idl`.
Created a `README.md` from `timezone.idl`. Going to create an
[Edify](https://github.com/bigeasy/edify) web site that will be the canonical
documentation. The essay on motivations, on dates and object orientation, will
be moved onto a separate page. It will get me to release faster to push that off
to the side and focus on documentation first.

Ran the Vim spell checker against `README.md` while I was at it.

Moved the `README` objectives into an **Objectives** section at the bottom of
the `README.md` page.

Did some writing, but stopped myself. Didn't feel like writing about motivations
any further. Motivations need to be a separate essay and not part of the
overview and introduction.

Almost completed #2.