Subtract years.
Subtraction of years was broken. I'd neglected to multiply the number of years
by the increment to negate the number of years if it is negative.

Created tests for year math, including a test for subtraction from a leap day to
a non leap day. Using the default JavaScript behavior, which is to adjust the
date to March 1st.

Closes #10.