Create Olson file tabulator.
Rather than poking around the Olson database for a zone and rule record pair as
I do now, I've created a program that creates a table for every timezone
containing a line for every ***seam*** in the timezone. A ***seam*** is any
point in time where the zone record and rule record from the Olson database
changes.

Every change to the UTC offset, daylight savings time savings, or timezone
abbreviation will be discovered.

Additional records may be discovered that indicate a shift in daylight savings
time rules, that do no actually change the UTC offset, daylight savings time
savings, or timezone abbreviation. These are transitions between zone records
where the set of daylight savings time rules change, but both sets of rules
agree on the daylight savings time savings. It indicates a political decision to
apply a different set of rules going forward from that date.

We detect all the seams so that we can test that we have interpreted all of
these transitions correctly.

This commit adds a program named `adjustments` that performs the above table
generation and emits a report. Going forward, the program will be generalized
and its primary function will become the timezone offset and abbreviation
resolution function in the heart of the `tz` function itself.

For now however, `adjustments` is merrily prattling away on `stdout`. It creates
a report used to confirm the output of the prarser.

The report will display UTC offset and timezone abbreviation before and after
each seam for all of the seams. (Currently, only the UTC offset is emitted.)

This report is fed to a shell program that compares the before and after output
with the the output generated by the system's `date`.