Created setClocks, added Docco.
Renamed `utc` to `posix` to bring it in line with the documentation. At the same
time, created a `setClocks` function to make the clock adjustment logic
consistent. Clocks are now set based on a `clock` property of an actualized rule
record or zone record. The `clock` property has a value of `posix`, `standard`
or `wallclock`.

Also, `extra` has become `carryover`, which is a better description of the
concept. We're carrying over a rule from the rules for the subsequent zone
record.

Added Docco to try to capture what's going on here. Some code tidy to delete
unused variables.