Make adjustments program a callback.
Converted the main method of the adjustments program to a accept a callback that
will build the table, instead of having it bulid the table itself.

This allows us to use the main method of the adjustments program to either build
the table, or in the future, to find the correct interval for a specific date.
However, at the moment, we do not have a way to terminate the traversal of a
timezone when the correct interval is found.