Positive offsets. Add all timezones.
Importing all of the timezones so that the program can test timezones outside of
North America. At this point, Detroit, Chicago and Menominee are working
correctly, so it's time to explore some of the old cities of Europe, looking for
those edge cases.

Already, a quick test against Amsterdam shows that I've got the formatting of
timezone offsets for positive offsets incorrect. The plus symbol must be there.