Zone records with `begins` instead of `until`.
When preparing the Olson files, we'll compile to intermediary format where the
`until` value of a zone record becomes a `begins` value of a subsequent record.
Here we create the new format from the existing format in `forward.js` prior to
walking the history. In the future it will be done in our `tz2json` program.

We will convert from Olson files to a JSON representation of the Olson files
using `tz2json`. We will convert from the JSON representation of the Olson files
to tables using `forward.js` or whatever we call it in the future.

Also, created a little script called sizes that will track the file sizes of the
`forward.js` file. I'm going to append the results to commit messages for now.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               10282/10.04  5132/05.01  3119/03.05  2166/02.12