Zone records with specific savings.
Some zone records include a daylight savings time offset in field that
ordinarily specifies a set of daylight savings rules. I've seen this Vilnius
where it appears that a new Vilnius government stomped on the Russian daylight
savings time rules. The zone record lasts only for that summer. Apparently, this
zone record saves the Olson file maintainers the hassle of creating a set of
Vilnius rules for that one summer.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               10703/10.45  5254/05.13  3245/03.17  2226/02.17