Correctly convert zone record clock.
The current intermediate format, totally broken, has two booleans to indicate
the clock, one for `utc` and one for `standard`.  Turns out, the `utc` flag is
ticked in addition the `standard` flag for clock transitions specified in
standard time. Changing the order of a few ternary operators fixes this.

This was causing problems parsing Vilnius, which led me to spend a day with the
`zic` compiler, trying to sort out a baffling new edge case. I was reading the
table below.

    Zone  Europe/Vilnius  1:41:16 -       LMT     1880
                          1:24:00 -       WMT     1917        # Warsaw Mean Time
                          1:35:36 -       KMT     1919 Oct 10 # Kaunas Mean Time
                          1:00    -       CET     1920 Jul 12
                          2:00    -       EET     1920 Oct  9
                          1:00    -       CET     1940 Aug  3
                          3:00    -       MSK     1941 Jun 24
                          1:00    C-Eur   CE%sT   1944 Aug
                          3:00    Russia  MSK/MSD 1991 Mar 31 2:00s
                          2:00    1:00    EEST    1991 Sep 29 2:00s
                          2:00    C-Eur   EE%sT   1998
                          2:00    -       EET     1998 Mar 29 1:00u
                          1:00    EU      CE%sT   1999 Oct 31 1:00u
                          2:00    -       EET     2003 Jan  1
                          2:00    EU      EE%sT

I'd read the interval ending on March 29th, 1998 incorrectly, forgetting that
rules are specified by the end of the interval, and thought that the `C-Eur`
rules were in effect. This meant that the `C-Eur` daylight savings time would
come into effect, but only for a few hours. This result was not being reported
by GNU `date` using the same files used by this library, compiled by `zic`.

My output was showing an anomaly, but that was owing to the clock transition
being misinterpreted as standard time instead of POSIX time. However, my
misreading, my simply forgetting how the file works, cost me four hours. Four
hours spent reading `zic.c`, trying to find the branch were a rule record clock
transition is rejected by cause it's, eh, close enough to zone record clock
transition.

I misread the file, until I realized that zic was reporting that the New Years
1998 clock transition had no subsequent rules, which caused me to look at the
file again, and it was ***obviously*** correct.

Then I went back to the `forward.js` file and by formatting the dates for each
record, I noticed that the March 29, 1998 clock transition was incorrectly
marked as being on the standard clock.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               10847/10.59  5313/05.19  3315/03.24  2251/02.20