Visit all zones going forward.
If no zone name is specified on the command line, visit every zone instead.
Catches errors and reports the name of zone rule set before rethrowing the error
for inspection. This is how we'll work through all of the zone rule sets and
assert that we can interpret them correctly.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               11114/10.85  5485/05.36  3413/03.33  2333/02.28