Clock transition range increased to a minute.
GNU `date` transitions timezones at the top of the minute, even if the time of
the transition is specified to the second. Updated `verify` to increase and
decrease the UTC time of the clock transition by a minute to detect if the zone
offset and zone abbreviation match their expected before and after values.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               11475/11.21  5649/05.52  3467/03.39  2349/02.29