Parse rules set for the 24th hour.
Amman springs forward at 24:00 on the last Thursday of March. This broke the
date calculation, because it uses the `Date` object to get the day of the week.
It created the whole date, with the hour, which pushed it forward one day.

Now searching for the last occurrence of day of the week uses a `Date` object
that is specified without the hours, minutes and seconds. The full `Date` object
with time and date is created after correct date is found.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               11747/11.47  5655/05.52  3634/03.55  2366/02.31