7-bit ASCII encoding of zoneinfo.
I created a 7-bit ASCII encoding of a compiled timezone rule set. I've found
that it does not have a significant savings over a JSON array of the same data,
after both formats have been gzipped.

Savings are extracted from JSON by eliminating significant digits. We do not
need millisecond accuracy, so nothing is lost by using seconds instead of
milliseconds. Most clock transitions fall on the top of the hour. We can divide
POSIX time clock transitions by 100 to get rid of two more significant zeros in
the JSON representation. Usually, it is only the first clock transition, from
sundial time, that occurred  at some odd minute. These odd minute clock
transitions express some values after decimal, making them longer. They are the
exception. All other values are shortened by the absence trailing zeros.

Further savings by storing the difference between POSIX time in seconds and
wall-clock time in seconds, instead of the full-wallclock time in seconds since
the epoch. Like the POSIX time value, the difference is then divided by 100 to
eliminate significant zeros, since the difference is usually a full hour.

We store the daylight savings in tens of minutes.

Abbreviations are placed at the head of the zoneinfo array, indexes are used in
the table to reference the indexes. Upon loading, the abbreviations are shifted
off the front of the array, then the abbreviation indexes are replaced with the
actual abbreviations.

The 7-bit ASCII encoding was a nice effort, but it didn't produce significant
savings after gzipping. It encoded values in 7-bit bytes, little-endian, placing
a sign bit in the most significant bit of the most significant byte. The raw
format is more compact than JSON, but compressed it is only about 8% smaller
than the compressed JSON.

The 7-bit ASCII requires a significant amount of extra code to decode. The
minified and gzipped size of the library is going to be a consideration for
adoption, so I'd prefer to push those bytes into the timezone data.

My rationale is not cynical. I do believe that too large a library is going to
invite unfavorable comparisons with other solutions, but I don't believe that
forgoing the 8% data savings here is a bait and switch.

My goal is to have Timezone be the same size to as a timezone challenged
library, with the same feature set, so that if you want to operate solely in
`UTC` with `en_US`, then you pay the same price in bandwidth.

To add timezone support, you need to add the timezone database. It doesn't make
sense to make it an all or nothing proposition. Any grouping is going to make
assumptions about usage with no usage to base those assumptions on. We may as
well serve rule sets a la carte.

When an application loads only a single timezone, they will probably load less
data with the JSON decoder, plus a single JSON file, than with the ASCII decoder
and a single ASCII file. If they are loading many timezones, then timezones are
an aspect of the application. As an aspect of the application, the bandwidth
costs of the timezone data are not going to be a barrier to adoption.

Also, as a barrier to adoption is a data file that a new programmer cannot
easily read. It will generate questions. It will require documentation.

Not to put too fine a point on it, I don't believe that considering barriers to
adoption is cynical. The developer of the timezone focused application would end
up wanting an 8% savings in his monthly bandwidth bill. If so, I'm happy to look
at the 7-bit ASCII encoding again. In the mean time, that developer will have
the shared support of a larger community, because the maintainer made decisions
to foster adoption of the code.

If someone does want to load a world of timezone data, however, it best for them
to pull down the by-continent JSON and generate the rule set tables as needed.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               16393/16.01  8623/08.42  4887/04.77  3376/03.30