Add first clock transition to JSON rule sets.
Added the first clock transition to the JSON rule sets.

I did a whitespace tidy. Probably should do those as separate commits. My
apologies. I'll do them separately in the future.

As of this commit there are two clock transitions for which the `forward.js`
output does not match the `verify` output.

`Asia/Damascus` on Fedora 16 at the time of commit is different from the IANA
data files. The records match when GNU `date` is run using the `zic` compiled
output from the IANA data files. For the future, `verify` will verify the output
of the Timezone zone rule set compiler against the output of `zic` for the same
data. To automate testing, the test will have to download the Time Zone Database
from the IANA along with the reference source code for the Time Zone Database
utilities, compile the utilities, and then compile `zoneinfo` tables using the
reference implementation of `zic`. The `verify` program will use use these
`zoneinfo` files with GNU `date` to verify the output of `forward.js`.

`Europe/Sofia` is off due to a zone record and a rule record overlapping.

    File                             Raw         Min         Zip     Min/Zip
    forward.js               17443/17.03  9310/09.09  5077/04.96  3569/03.49