Pull `7-bit.js`, `display.js` from `forward.js`.
Extracted the 7-bit ASCII encoding into `7-bit.js`. Extracted the display of the
clock transitions and zone record seams to `display.js`.

The `7-bits.js` and `forward.js` programs will continue to verify their encoding
after encoding by decoding after encoding and comparing the decoded table with
the original table.

The `display.js` program is used to verify that the output of `transitions.js`
is correct, by creating a plain text listing of all clock transitions that can
be compared against the output of GNU `date`, as we've been doing with the
`verify` program.

    File                             Raw         Min         Zip     Min/Zip
    forward.js                2702/02.64  1541/01.50  1007/00.98   812/00.79
    transitions.js            9261/09.04  4160/04.06  2843/02.78  1731/01.69
    7-bit.js                  4555/04.45  2849/02.78  1591/01.55  1282/01.25
    display.js                1431/01.40   774/00.76   645/00.63   482/00.47
    timezones/asia.js      131087/128.01 60231/58.82  7420/07.25  6351/06.20