Redo zone record preprocessing.
Moving away from having legible JSON in the Olson as JSON files, starting with
the zone records. Instead of an ISO 8601 string from the until property, it's
seconds since the epoch. Using `false` for the maximum zone record, instead of
just leaving the property out of the record. This is because I intend to move to
arrays instead of objects for zone records.

Converted the silly `utc` and `standard` booleans to a `clock` property. This is
going to be single integer to save us the quote characters before we're done.

Converting the offset to seconds since the epoch. No savings in JSON size
because, for the most part, the offsets are specified in six characters as
strings.

I'm only going to commit the changes to `tz2json.coffee` and `transitions.js`
for now. I'll commit the changes to the Olson as JSON files in one final commit
before merging this branch back into the `tabular` branch.

I'll be using `asia.js` as a measure of progress toward a smaller Olson as JSON
file format.

    File                             Raw         Min         Zip     Min/Zip
    timezones/asia.js      130174/127.12 58705/57.33  7520/07.34  6482/06.33

Notice how the raw and minified JSON is shrinking, but the zipped and minified
and zipped JSON is growing.

Date strings might be more efficient than seconds since the epoch. At some point,
we may look to using the `tz` function itself to parse the Olson to JSON files.
Any time we're messing around with these files, the `tz` function will be close
by. Invoking it is not verbose.

Date are an option if they compress better.