Redo rule record preprocessing.
Setting the clock in the conversion and removing its detection from
actualization. Adding the clock increased the size of the Olson as JSON file.

Removed the unused `type` property.

Converted `time` to minutes. Simplifies `actualize` slightly. It is almost
always the top of an hour in the first few hours of the day, so it can't be any
less repetitive than the strings.

Converted `save` to minutes. Savings in the raw file size of `transitions.js`,
but no savings, none at all, after minify and gzip. Saved 17 bytes in the Olson
as JSON file.

Converted `rules` to minutes When `rules` is a savings. Thereafter, I was able
to remove `parseOffset` from `transitions.js` saving 114 bytes minified and
gzipped.

    File                             Raw         Min         Zip     Min/Zip
    transitions.js            7692/07.51  3418/03.34  2392/02.34  1449/01.42
    timezones/asia.js      130742/127.68 59273/57.88  7535/07.36  6429/06.28