Sort precompiler files into directories.
Moved the new precompiler files split out from `forward.js` to the `lib` and
`bin` directories. Going forward we'll update `7-bit.js` with any further
changes to `transitions.js`. If we ever decide that the 7-bit ASCII encoding is
the way to go, we'll have ready and waiting. There's no real need, at this
juncture, to delete the file.

    File                             Raw         Min         Zip     Min/Zip
    lib/transitions.js        7147/06.98  3171/03.10  2279/02.23  1376/01.34
    bin/7-bit.js              4562/04.46  2856/02.79  1595/01.56  1286/01.26
    bin/forward.js            2708/02.64  1547/01.51  1010/00.99   816/00.80
    bin/display.js            1445/01.41   788/00.77   647/00.63   486/00.47