No longer going backwards.
Deleted `adjustments.coffee`. No longer going backwards. Certainly can't go
backwards through the Olson files as they are. They would need to be converted
into a format that permits going backwards.

Maybe at some point in the future I'll once again consider a format that
calculates as needed. It would be the only way to get the size of the rule set
tables smaller, to eliminate entries.

Offhand, with all that I know now, I can say that it would be possible to
convert the transitions to wall-clock time within each zone record, for a
particular zone. The problem with walking backwards was calculating wall-clock
time given a potpourri of rules. If our precompiler converts those rules to
wall-clock for us, then we can scan backwards. We need for all the times to be
set to the same clock for a zone entry.

Our precompiler could add to a zone, only those rules that are applicable, not
the full set of rules. It can adjust the rules to simplify their selection. I
did learn from `adjustments.coffee` that rules are sorted by `from`, but for
traveling backwards through them they need to be sorted by `to`.

That is another deep dive, should I ever choose to dive into this problem again.

We will have rule actualization in `tz` anyway, to actualize rules for future
dates. If we can build a backwards scan using that actualization code, then
we've paid for the compress JSON rule set tables already.

If it does require a significantly larger `tz`, it is a dubious savings. It is
important that `tz` itself be small, so it can be adopted for timezone ignorant
uses. To grow it in order save some bytes in the JSON rule set tables is not a
good trade-off. Slightly smaller JSON rule set tables are unlikely to unlock a
whole new initialization pattern.