Correctly calculate standard time from wall-clock.
When calculating standard time from wall-clock time, I was adding the day light
savings instead of subtracting it. This created an incorrect clock transition
for `Europe/Sofia`, the only rule set effected.

Now that this this is fixed, all of the clock transitions can be verified using
GNU `date` against the `zic` compiled IANA Timezone Database.

Closes #15.