Update Travis-CI build for `Makefile`.
Updated `.travis.yml` to invoke `make` as part of the continuous integration
build. This creates the `zic` compiled IANA Timezone Database on the Travis-CI
worker machine.

Removed `bin/verify` from the `Makefile` because it was causing Travis-CI to
timeout. Travis-CI was unhappy about the time it took to run the
`before_install` programs. Can't see the exact message at the moment, because
Travis-CI is overloaded, probably due to RailsConf.

Later on, when we've converted `bin/verify` to work with `proof`, we'll see if
Travis CI will give us long enough to run all the `bin/verify` tests as part of
the actual build.

Updated for latest CoffeeScript. CoffeeScript 1.3.1 did not like my `do`
incantations that are intended to select the correct object for export, either
`module.exports` for `require`, or `window` for the browser.

It also thought that `02` for the hour of the `moonwalk` in our `proof` test
harness was octal and malformed according to the new CoffeeScript syntax for
octals. That `02` is meant to decimal, so I got rid of the leading zero.

There were CoffeeScript 1.3.1 problems in Twinkie as well. They've been patched
and Timezone uses a latest Twinkie.