Remove locale aware date parsing.
No natural language processing what-so-ever. Out of scope for this library.
Where do you draw the line with natural language processing? The slightest
amount sets expectations for more.

Without it, I can claim to support ISO 8601 and RFC 3339 dates. (I'm considering
splitting RFC 822 out and into a conversion funtion.) Without it, I can point to
teh standards and say that string dates are supported solely to support

 * a time format that includes timezone offsets,
 * a time format that allows for converstion from wall-clock to UTC.

With it, I have to draw a line on fuzzy date parsing, then defend that line.

Without it, this is date math library.

With it, this is a localization library.

In fact, I could make this library much smaller, if formatting converted from
ISO 8601 to a formatted date. The formatting could be separate from the date
math. It is bulky to support all of UNIX date, plus locales.

I've removed the slash delimited date parsing. The lines of code after the slash
delimited date parsing have been nothing but `TODO` comments since this project
began. This was a requirement informed by naive notions of what a date parsing
library should be. There will be more veribage about the scope of parsing in the
`synopsis.js` documentation, so I'll save it for there.

Also, found a couple of calls to `Date.UTC.apply` passing `null` as the context
object. Changed that to `Date.UTC`.